add dictionary handlers to validator
