+++ src/Handler/Dictionary/dict/de.php
