+++ src/Handler/Dictionary/dict/es.php
