+++ src/Handler/Dictionary/dict/sk.php
